Add display rotation feature
