+++ src/mini_12864_module.h
@@ -1,8 +1,8 @@
-#ifndef ROTARY_BUTTON_H_
-#define ROTARY_BUTTON_H_
+#ifndef MINI_12864_MODULE_H_
+#define MINI_12864_MODULE_H_
 
 
-#define EEPROM_ROTARY_BUTTON_DATA_REV           1
+#define EEPROM_MINI_12864_MODULE_DATA_REV           2
 
 #include "http_rest.h"
 
@@ -18,12 +18,19 @@ typedef enum {
     OVERRIDE_FROM_REST,
 } ButtonEncoderEvent_t;
 
+typedef enum {
+    DISPLAY_ROTATION_0 = 0,
+    DISPLAY_ROTATION_90 = 1,
+    DISPLAY_ROTATION_180 = 2,
+    DISPLAY_ROTATION_270 = 3,
+} display_rotation_t;
+
 
 typedef struct {
-    uint32_t rotary_button_data_rev;
+    uint32_t data_rev;
     bool inverted_encoder_direction;
-
-} rotary_button_config_t;
+    display_rotation_t display_rotation;
+} mini_12864_module_config_t;
 
 
 #ifdef __cplusplus
@@ -35,9 +42,11 @@ extern "C" {
  * Wait for button encoder input. 
 */
 ButtonEncoderEvent_t button_wait_for_input(bool block);
-bool button_init();
+bool mini_12864_module_init(void);
+void button_init(void);
+void display_init(void);
 bool http_rest_button_control(struct fs_file *file, int num_params, char *params[], char *values[]);
-bool http_rest_button_config(struct fs_file *file, int num_params, char *params[], char *values[]);
+bool http_rest_mini_12864_module_config(struct fs_file *file, int num_params, char *params[], char *values[]);
 bool button_config_save();
 
 #ifdef __cplusplus
@@ -46,4 +55,4 @@ bool button_config_save();
 
 
 
-#endif  // ROTARY_BUTTON_H_
+#endif  // MINI_12864_MODULE_H_
